requirements.txt
aiohttp
aiofiles
pandas
livekit-api